pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llrq"
version = "0.0.2"
description = "Log-Linear Reaction Quotient Dynamics: Tractable framework for chemical reaction networks"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Steven Diamond", email = "steven@gridmatic.com"},
]
keywords = ["reaction networks", "systems biology", "SBML", "chemical kinetics", "reaction quotient"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Mathematics",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "python-libsbml>=5.19.0",
    "tellurium>=2.2.0",
]

[project.urls]
Homepage = "https://github.com/reactionquotient/llrq"
Repository = "https://github.com/reactionquotient/llrq"
Issues = "https://github.com/reactionquotient/llrq/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 127
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 127

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=5.0",
    "mypy>=1.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "psutil>=5.9.0",
]
